backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pdfplumber
Pillow
requests
